Update bms-network-setup.py - some more refactoring
Fixed some Exception stylings - it would be best to define custom Exceptions.

added ```startswith``` and ```endswith``` for better readability

removed several unused parameters in functions:
Unused parameters are misleading. Whatever the value passed to such parameters is, the behavior will be the same.

rewrote the code snippet that checks if jval is not None to a more concise way to express this

Defined a constant instead of duplicating this literal ".yaml" 3 times.